requirements.txt
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Flask==3.1.0
gtfs-realtime-bindings==1.0.0
gunicorn==23.0.0
humanize==4.11.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
packaging==24.2
protobuf==5.29.3
pycron==3.1.2
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
urllib3==2.3.0
Werkzeug==3.1.3
